fix: move test_participants_for_meeting to meeting tests
